Bugfix: Fix learning rate rounding error
- Modified gradDesc method declaration in ConvLayer.h
- Ensure learningRate parameter is double, not int
- Prevent low learning rates rounding to zero and propagating errors throughout network

+++ include/ConvLayer.h
@@ -35,7 +35,7 @@ public:
 
     void backProp(Eigen::MatrixXf nextLayerW, Eigen::MatrixXf nextLayerDz, Eigen::MatrixXf layerInputMat, bool prevLayerConv);
     
-    void gradDesc(int learningRate);
+    void gradDesc(double learningRate);
     
     Eigen::MatrixXf getZ();
     Eigen::MatrixXf getA();
